Rename AsyncImage module and update dependencies for design system integration
