+++ src/build-logic/src/main/kotlin/plugins/popcorngp-setup-plugin.gradle.kts
@@ -45,9 +45,7 @@ popcornGuineapigParentConfig {
         PopcornChildConfiguration(
             moduleNameRegex = ":designsystem",
             rules = listOf(
-                JustWithRule(
-                    justWith = listOf("media"),
-                ),
+                NoDependencyRule(),
             ),
         ),
     )
